Add _get_clock_path_delay return check
Sometimes, _get_clock_path_delay method may not find the clock path.
In this case, we should add a return check to see whether it find
the clock path or not.
In some cases, we should insure that all ff nodes can find its clock.
If not, raise an error.